fix cell port behavior
we should not be expecting all app ports to be 8080 and should instead
be looking for the first port that is not used for SSH (2222).
this normally occurs for docker apps, where you can pick any port.
there's still some uncertainty around whether we can have more than one
port for this.

[#156766522]

Signed-off-by: Zachary Gershman <zgershman@pivotal.io>